requirements.txt
numpy
Pillow
PyYAML
tqdm